+++ src/test/java/kz/trankwilizator/tafl/entity/persist/PermanentUserEntityPersistTest.java
@@ -1,8 +1,9 @@
-package kz.trankwilizator.tafl.entity.user.permanent;
+package kz.trankwilizator.tafl.entity.persist;
 
 import jakarta.persistence.PersistenceException;
 import kz.trankwilizator.tafl.dao.user.PermanentUserRepository;
 import kz.trankwilizator.tafl.entity.user.PermanentUser;
+import kz.trankwilizator.tafl.entity.config.user.PermanentUserEntityTestConfig;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
